extract version resolver to shared utils
